backend/requirements.txt
annotated-types==0.7.0
async-timeout==5.0.1
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
redis==5.2.1
SQLAlchemy==2.0.38
typing-inspection==0.4.1
typing_extensions==4.12.2
Werkzeug==3.1.3
